Added a working TCP server that serves to host a simulator of switches. Server is written in Rust and a seperate Python script is created to interact with the server. The server tries it's best to emulate the protocol of what the switches are supposed to based on interaction with LISC and addressable protocols.
The emulated 'switches' on server only know two commands at the moment:

* GoIdle
* GetStatus

The emulated switches automatically begin broadcasting their id
The server only holds 4 switches at the moment.

The purpose of this server is to test integrity and code of the original GUI
that handles all the protocol and logic used when interacting with real switches.