A complete, however bare, switch simulator running on a seperate server. Currently have 4 working switches that respond with 2 basic commands:
* GetStatus and GoInactive
Right now the server is very rail-roady in that it iterates over the 4
built-in virtual switches and the GUI simulator sends it through the stages of:

Broadcasting -> SendStatus -> GoIdle

Repeating 4 times.

TODO:
Use the simulation window as a proxy to forward data from server to the
main GUI where main GUI code can be used to parse information from server.
This can act as a test to withstand the integrity of the code..